refactor(dependencies): moved to more according package
